fix utest

+++ python/paddle/fluid/tests/unittests/test_math_op_patch.py
@@ -192,12 +192,12 @@ class TestMathOpPatches(unittest.TestCase):
         b = a / 7
         place = fluid.CPUPlace()
         exe = fluid.Executor(place)
-        a_np = numpy.array([3, 4, 10, 14, 9, 18]).astype('int64')
+        a_np = numpy.array([3, 4, 10, 14, 9, 18])
         b_np, = exe.run(fluid.default_main_program(),
                         feed={"a": a_np},
                         fetch_list=[b])
 
-        b_np_actual = (a_np / 7).astype('int64')
+        b_np_actual = (a_np / 7)
         self.assertTrue(numpy.array_equal(b_np, b_np_actual))
 
     @prog_scope()
